+++ internal/ops/chops/chops.go
@@ -77,6 +77,8 @@ func lakeMutatorVerb(verb string) (func([]string) error, bool) {
 		return chContractLedgersBackfill, true
 	case "ch-instance-backfill":
 		return chInstanceBackfill, true
+	case "ch-census-rollup":
+		return chCensusRollup, true
 	case "ch-cap67-movements":
 		return chCap67Movements, true
 	case "ch-holders-rollup":